requirements.txt
Flask==2.3.2
flask-cors==5.0.1
opencv-python==4.8.0.76
torch==2.6.0
transformers==4.31.0
Pillow==10.0.0
numpy==1.25.2
setuptools==67.6.1